Added support for reading .cls files
Added support for reading .cls files, the convention is that in the same
directory as the .trace file is the .cls file, they only differ by the
filename suffix (.trace vs .cls). All the classes will be loaded into
the classTable which is part of the Memory Manager. In order to retrieve
the class name query getClassName(classID) from the memory manager, the
string "CLASS_TABLE_NOT_LOADED" indicates that the class table is not
loaded. If the value is out of bounds the string "OUT_OF_BOUNDS" will be
returned