Fixed a bug in the real allocator
Fixed a bug in the real allocator which moved the object to the same
space. Usually if an object survives it has to be moved to the new
space, however the old implementation moved it within the old space.
This is fixed now. The real allocator itself is still not working
correctly though, working on it