Changed the file output
Changed the file output so that it creates a log file in the same
directory as the tracefile. Removed the gcTimes.log as it is now part
of the log file. Introduced a toggle (-f) in order to force a GC after
every step, which can be used in order to verify your GC, the filename
for the log will have a ‘Forced” appended to it. Currently the bug is
that for the traversal collector the heap size is displayed incorrectly
as we are using a split heap for that collector. Working on that now.