Fixed tons of bugs, more testing needed
Fixed tons of bugs in the simulated allocator to make it work with the
traversal GC. markSweep GC still works with it. Broke the real allocator
while fixing the simulated allocator. Proceed with caution, as both
allocators might not work with your implementation. Will do more testing
and will fix the real allocator