Revert "Extracted class table management into separate class."
This reverts commit 101187ffdbdfbfa58b1c64c009c46500b514ffb4.

This class was extracted incorrectly and has led to a subtle bug when performing
simulations using the simulated allocator. "Free Heap" is being reported as
negative values in the *.log summary file. These negative values are only being
reported when a *.cls file is included with the simulation run; if the *.cls
file is ommitted, the "Free Heap" is being reported (seemingly) correctly.
Additionally, the reported negative values only appear when compiled and run on
an x86 architecture (GMC in this case); when compiled and run on a ppc
(Lincoln), the reported values are as expected.

Reverting this change as a quick-fix for the bug.