Multithreaded for I/O, fixed clock, startup
- extracted the I/O with the tracefile into its own thread (resulted in
a speedup >3)
- cannot use clock() for multithreaded execution, as it retrieves CPU
time and multithreaded execution uses more CPU time in less time.
Exchanged it for clock_gettime() which gives the wall time
- now a human readable format for heap sizes can be used, e.g. the
toggle "-h 10m" or "-h 10M" or "-h 10 MB" initializes a heap with the
size of 10 MB (2^10 used as base). No suffix results in the number
interpreted as bytes, just as before

!!! NEEDS TESTING BEFORE PUSHING BACK TO MASTER !!!