Fix. Simulator always assumed split heap.
When initializing the allocator, the heap was always being treated as a split
heap. This was correct for the traversal collector (which performs copying
between 'new' and 'old' space), but was incorrect for other policies. This led
to only half of the heap being used during a simulation run and led to seemingly
excessive heap requirements.

Fixed so that the other GC policies receive appropriate heap sizes.