Refactor. Changed the way that 'Heap Used' and 'Heap Free' columns get generated.
Previously, heap usage statistics were being maintained as objects were being
allocated and freed. This is efficient, but causes problems as the allocation
and collection policies are extended. Anytime someone creates a new way to
allocate or free space, they must remember to increment/decrement the statistics
accordingly.

I've re-implemented these methods to generate the statistic at the time the
method is called. This is less efficient (it's linear in the size of the heap),
but it is stand-alone. Further enhancements / modifications should not skew the
statics being returned. I believe that this is a fair trade-off.