Fixed trigger for 'shift' initiated collection.
The markSweep policy enters an infinite loop if:
	a) shifting is enabled
	b) an inadequate heap size is specified.

This is caused because shifting is intended to free-up heap space by calling
collections until objects become 'old enough' to be promoted. Since the
markSweep policy does not include a promotion mechanism, when this situation
arises, the collection is called repeatedly.

This issue has been fixed by only triggering a shift initiated GC is the number
of generations is greater than 1.